Move remaining testing classes into tests dir
