+++ tests/Google/Ads/GoogleAds/Lib/google_ads_php.ini
